requirements.txt
progressbar2==3.38.0
matplotlib==2.2.2
numpy==1.15.0